requirements.txt
discord.py==2.4.0
pillow==11.0.0
tqdm==4.67.1
